+++ tests/test_result/test_helper_functions/test_safe.py
@@ -2,8 +2,7 @@
 
 import pytest
 
-from returns.functions import safe
-from returns.result import Failure, Success
+from returns.result import Failure, Success, safe
 
 
 @safe